Move XCFunctional to a separate mrdft folder
The MRDFT module is now ready to break loose
and evolve on its own, if desireable.
