+++ src/CMakeLists.txt
@@ -13,6 +13,7 @@ add_subdirectory(chemistry)
 add_subdirectory(analyticfunctions)
 add_subdirectory(qmfunctions)
 add_subdirectory(qmoperators)
+add_subdirectory(mrdft)
 add_subdirectory(scf_solver)
 add_subdirectory(initial_guess)
 add_subdirectory(utils)
