+++ src/qmoperators/two_electron/CMakeLists.txt
@@ -2,6 +2,5 @@ target_sources(mrchem PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/CoulombPotential.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/ExchangePotential.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/FockOperator.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/XCFunctional.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/XCPotential.cpp
     )
